fix store.delete
